// libs/features/feature360/src/lib/models/establishments/establishment-molprofile.ts
import { BilingualText } from '@gosi-ui/core';

export class EstablishmentMOLProfile {
  molestid: number = undefined;
  molestofficeid: number = undefined;
  moluniid: number = undefined;
  molofficeid: number = undefined;
  establishmentname: BilingualText = new BilingualText();
  legalentity: number = undefined;
  establishmentstatus: number = undefined;
  main: number = undefined;
  ownerninum: number = undefined;
  ownerfirstname: BilingualText = new BilingualText();
  ownersecondname: BilingualText = new BilingualText();
  ownerthirdname: BilingualText = new BilingualText();
  ownerlastname: BilingualText = new BilingualText();
  recruitmentnumber: number = undefined;
  LegalEntityArabic: string = undefined;
  LegalEntityEnglish: string = undefined;
  ChceckIFMAINArabic: string = undefined;
  ChceckIFMAINEngalish: string = undefined;
  ESTABLISHMENTSTATUSArabic: string = undefined;
  ESTABLISHMENTSTATUSEnglish: string = undefined;
}