// libs/features/feature360/src/lib/models/establishments/rased-advanced-search-details.ts
import { BilingualText, GosiCalendar } from '@gosi-ui/core';

export class RasedAdvancedSearchDetails {
  id: BilingualText = new BilingualText();

  inspectionid: number = undefined;
  establishmentregistrationno: BilingualText = new BilingualText();
  establishmentname: BilingualText = new BilingualText();
  inspectionrequestno: string = undefined;
  requesteddocumentyn: BilingualText = new BilingualText();
  inspectionfieldactivityid: number = undefined;
  fieldactivityno: BilingualText = new BilingualText();
  reportid: BilingualText = new BilingualText();
  activitydate: BilingualText = new BilingualText();
  claimno: BilingualText = new BilingualText();
  initiatedreason: BilingualText = new BilingualText();
  originatedsystem: BilingualText = new BilingualText();
  priority: BilingualText = new BilingualText();
  status: BilingualText = new BilingualText();
  visittype: BilingualText = new BilingualText();
  createdby: BilingualText = new BilingualText();
  createdon: GosiCalendar = new GosiCalendar();
  inspectiontype: BilingualText = new BilingualText();
  validatordecision: BilingualText = new BilingualText();
  inspectorname: BilingualText = new BilingualText();
  visitaddressline: BilingualText = new BilingualText();
  establishmentvisittype: BilingualText = new BilingualText();
  description: BilingualText = new BilingualText();
  bpmusertaskbean: BilingualText = new BilingualText();
  reportstatus: BilingualText = new BilingualText();
  createdbyname: BilingualText = new BilingualText();
  contributorregno: BilingualText = new BilingualText();
  contributorname: BilingualText = new BilingualText();
  region: BilingualText = new BilingualText();
  addperiodyn: BilingualText = new BilingualText();
  inspectiondate: BilingualText = new BilingualText();
  contributordecisionstatus: BilingualText = new BilingualText();
  pendingwith: BilingualText = new BilingualText();
}